requirements.txt
pyglet~=2.0.17
colorama~=0.4.6
pyfiglet~=1.0.2
speedtest-cli~=2.1.3
lorem_text~=2.1